+++ app/Http/Controllers/UserController.php
@@ -5,6 +5,8 @@ namespace App\Http\Controllers;
 use App\Models\User;
 use App\Traits\ApiResponder;
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Response;
 use Illuminate\Support\Facades\Hash;
 use Illuminate\Support\Facades\Validator;
 
@@ -12,6 +14,11 @@ class UserController extends Controller
 {
     use ApiResponder;
     
+    public function __construct()
+    {
+        // $this->middleware('auth'); 
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -20,6 +27,7 @@ class UserController extends Controller
     public function index()
     {
         try{
+            $this->authorize('user-module');
             $user = User::all();
             return ApiResponder::successResponse($user);
         }catch(\Exception $e){
@@ -36,9 +44,11 @@ class UserController extends Controller
     public function store(Request $request)
     {
         try{
+            $this->authorize('user-module');
             $data = $request->all();
             $validator = Validator::make(request()->all(), [
                 'username' => 'required|string|max:255|unique:users',
+                'role_id' => 'required|max:255',
                 'password' => 'required|string|min:6|confirmed',
             ]);
 
@@ -49,10 +59,12 @@ class UserController extends Controller
             $user = User::create([
                 'name' => ($data['name']) ? $data['name'] : '',
                 'username' => $data['username'],
+                'role_id' => $data['role_id'],
                 'password' => Hash::make($data['password']),
             ]);
 
             return $this->successResponse($user, 201);
+            
         }catch(\Exception $e){
             return $this->errorResponse($e->getMessage(),$e->getCode());
         }
@@ -69,6 +81,7 @@ class UserController extends Controller
     public function update(Request $request, User $user)
     {
         try{
+            $this->authorize('user-module');
             $validator = Validator::make(request()->all(), [
                 'name' => 'required|string|max:255',
                 'username' => 'required|string|max:255|unique:users,username,'.$user->id,
@@ -95,6 +108,7 @@ class UserController extends Controller
     public function show(User $user)
     {
         try{
+            $this->authorize('user-module');
             return $this->successResponse($user);
         }catch(\Exception $e){
             return $this->errorResponse($e->getMessage(),$e->getCode());
@@ -110,6 +124,7 @@ class UserController extends Controller
     public function destroy(User $user)
     {
         try{
+            $this->authorize('user-module');
             $user->delete();
             return $this->successResponse('User Deleted',200);
         }catch(\Exception $e){
